% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{clim_distweight}
\alias{clim_distweight}
\title{climdata distance weight
Apply distance weighting to point climate data extracted from era5 grid}
\usage{
clim_distweight(lon, lat, climr, tme)
}
\arguments{
\item{lon}{decimal longitude}

\item{lat}{decimal latitude}

\item{climr}{list of spatrasters of climate variables (output of mcera5::extract_clima)}

\item{tme}{POSIX vector of hourly times represented in climr}
}
\value{
dataframe of climate variables for input to micropoint::runmicropoint
}
\description{
climdata distance weight
Apply distance weighting to point climate data extracted from era5 grid
}
