% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.bbox_to_char}
\alias{.bbox_to_char}
\title{Convert a bounding box to a character vector}
\usage{
.bbox_to_char(bbox, coord_order)
}
\arguments{
\item{bbox}{an object that has a bounding box. Currently restricted to a SpatVector or SpatRaster, but will add functionality for other objects}

\item{coord_order}{the order in which coordinates should be output (e.g., "xmin,xmax,ymin,ymax") with no white spaces}
}
\description{
Convert a bounding box to a character vector
}
