% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.get_grid}
\alias{.get_grid}
\title{get_grid:Make a global template for downloading gridded data}
\usage{
.get_grid(crs, tilesize)
}
\arguments{
\item{crs}{character; coordinate reference system for the grid (default: "epsg:4326")}

\item{tilesize}{numeric; tilesize for the grid (180 and 90 should be evenly divisible by tile size)}
}
\description{
get_grid:Make a global template for downloading gridded data
}
