% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.ground_point}
\alias{.ground_point}
\title{Ground_point
Make groundparams object for input into micropoint::runmicropoint}
\usage{
.ground_point(soil, asp_p, slp_p, lon, lat)
}
\arguments{
\item{soil}{3 layer spatraster of clay, sand, and silt content (from soilgrids)}

\item{asp_p}{aspect (degrees) at lon, lat}

\item{slp_p}{slope (degrees) at lon, lat}

\item{lon}{longitude decimal degrees}

\item{lat}{latitude decimal degrees}
}
\description{
Ground_point
Make groundparams object for input into micropoint::runmicropoint
}
