% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.pai_vertprofile}
\alias{.pai_vertprofile}
\title{PAI vertical profile
Calculate pai vertical profile from gedi data
Values add up to total PAI (for input into micropoint paii argument)}
\usage{
.pai_vertprofile(pai_z, pai, h)
}
\arguments{
\item{pai_z}{vector of PAI values at height intervals from 0 to top of canopy (as in gedi data)}

\item{h}{height of canopy}
}
\value{
PAI profile at 1 m intervals from ground to top of canopy. Values add up to total PAI in gedi data
}
\description{
PAI vertical profile
Calculate pai vertical profile from gedi data
Values add up to total PAI (for input into micropoint paii argument)
}
