% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.reproj_bigr}
\alias{.reproj_bigr}
\title{Reproject big raster
Crop and reproject large raster to specified extent and projection of extent}
\usage{
.reproj_bigr(r, proj_e, e, method)
}
\arguments{
\item{r}{spatraster to crop and reproject}

\item{proj_e}{projection of spatextent}

\item{e}{spatextent to project to}

\item{method}{method for reprojection (see terra::project method for options)}
}
\description{
Reproject big raster
Crop and reproject large raster to specified extent and projection of extent
}
