% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{era_process_splinedtiles}
\alias{era_process_splinedtiles}
\title{process pre-downloaded and splined era5 data}
\usage{
era_process_splinedtiles(path, lsmpath, r, year, resampleout = FALSE)
}
\arguments{
\item{path}{path to directory with nc files (located in /orange/scheffers/dklinges/soilTemp/SoilTemp-big-data/climate/era5/global/YYYY/era5_lon_lat/)}

\item{lsmpath}{path to land-sea-mask}

\item{r}{spatRaster template for output}

\item{year}{year of climate data to obtain}

\item{resampleout}{logical; if TRUE resamples era5 to template. default = F}
}
\description{
process pre-downloaded and splined era5 data
}
