% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{extract_clima_2}
\alias{extract_clima_2}
\title{Process era5 data using extract_clima}
\usage{
extract_clima_2(
  nc,
  long_min,
  long_max,
  lat_min,
  lat_max,
  start_time,
  end_time,
  land_sea_mask,
  dtr_cor = TRUE,
  dtr_cor_fac = 1.285,
  format = "microclimf"
)
}
\arguments{
\item{nc}{character vector containing the path to the nc file. Use the
\code{build_era5_request} and \code{request_era5} functions to acquire an nc file with
the correct set of variables. Data within nc file must span the period
defined by start_time and end_time.}

\item{long_min}{minimum longitude of the grid for which data are required (decimal}

\item{long_max}{maximum longitude of the grid for which data are required (decimal
degrees, -ve west of Greenwich Meridian).}

\item{lat_min}{minimum latitude of the grid for which data are required (decimal}

\item{lat_max}{maximum latitude of the grid for which data are required (decimal
degrees, -ve south of the equator).}

\item{start_time}{a POSIXlt or POSIXct object indicating the first day or hour
for which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{end_time}{a POSIXlt or POSIXct object indicating the last day or hour for
which data are required. Encouraged to specify desired timezone as UTC (ERA5
data are in UTC by default), but any timezone is accepted.}

\item{land_sea_mask}{spatraster; era5 lsm layer}

\item{dtr_cor}{logical value indicating whether to apply a diurnal temperature
range correction to air temperature values. Default = \code{TRUE}.}

\item{dtr_cor_fac}{numeric value to be used in the diurnal temperature range
correction. Default = 1.285, based on calibration against UK Met Office
observations.}

\item{format}{specifies what microclimate package extracted climate data will
be used for. Data will be formatted accordingly. Default is "microclimf".
Options: "microclima", "NicheMapR", "microclimc", "microclimf", "micropoint".
Note: of these options, only "microclimf" accepts as input an array of climate
variables. For all other models you will need to iterate through each spatial
point to run the model}
}
\description{
Forked from mcera5::extract_clima. Adjustments include, faster cropping to extent
and addtion of land-sea mask separately from the era5.nc file
}
