% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{gedi_finder}
\alias{gedi_finder}
\title{Find gedi orbits}
\usage{
gedi_finder(product, bbox)
}
\arguments{
\item{product}{can be  'GEDI01_B.002', 'GEDI02_A.002', 'GEDI02_B.002'}

\item{bbox}{bounding box coords in LL Longitude, LL Latitude, UR Longitude, UR Latitude format
bbox should be a character "LLlong,LLlat,URlong,URlat" coordinates in WGS84}
}
\value{
list of granules with shots in bbox
}
\description{
Define Function to Query CMR
The function returns a list of links to download the files of GEDI transects intersecting the bounding box
sub-orbit V2 granules directly from the LP DAAC's Data Pool.
}
