% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{get_climate}
\alias{get_climate}
\title{Download and process climate data
use functions from microclimdata package}
\usage{
get_climate(
  source,
  r,
  tme,
  creds,
  tilepath,
  lsmpath,
  out = "grid",
  pathout,
  file_prefix,
  clean = F
)
}
\arguments{
\item{source}{character; source to get climate data. Accepts
era5
era5-hipergator-splined: this gets data from Dave's splined tiles on hipergator}

\item{r}{a SpatRaster object used as a template for downloading data with an EPSG code defined. Should be in a projected crs.}

\item{tme}{POSIXct vector of hourly time intervals to download data (must be in UTC)}

\item{creds}{dataframe with credentials as specified in microclimdata}

\item{tilepath}{if source = era5-hipergator-splined, tilepath = path to splined  era5 tiles on hipergator}

\item{lsmpath}{if source = era5-hipergator-splined, provide path to era5 land-sea-mask}

\item{out}{one of 'grid' or 'point' specifying whether to return a list of arrays of climate data or a data.frame for a single point location.}

\item{pathout}{directory to store data. Separate folders will be created for raw and processed data. raw downloaded data.}

\item{file_prefix}{prefix used for file name}

\item{clean}{logical; if T, deletes original downloads, if F, keeps original downloads}
}
\description{
Download and process climate data
use functions from microclimdata package
}
