% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_era5}
\alias{get_era5}
\title{Download ERA5 for area}
\usage{
get_era5(st_time, en_time, uid, file_prefix, file_path, e)
}
\arguments{
\item{st_time}{start time of download, character date in format YYYY-MM-DD (e.g., "2023-01-01")}

\item{en_time}{end time of download, character date in format YYYY-MM-DD (e.g., "2023-12-31)}

\item{uid}{uid for climate data store (should be your email for your account)}

\item{file_prefix}{prefix for .nc file}

\item{file_path}{path for .nc file}

\item{e}{spat extent to define spatial extent of the download}
}
\value{
NA downloads and saves data
}
\description{
Download ERA5 climate data using mcera5. Will not download if file is already present.
cds access token must be set prior to running this function. See mcera5 vignette for details
}
