% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_gedi_tiles}
\alias{get_gedi_tiles}
\title{Download GEDI tiles}
\usage{
get_gedi_tiles(e, fpath, unzip = F, download = T)
}
\arguments{
\item{e}{spatraster, spatvector, or spatextent}

\item{fpath}{file path for downloaded GEDI tiles}

\item{unzip}{logical; should the downloaded file be unzipped warning if true, file sizes will be large}

\item{download}{logical (default TRUE) if TRUE will download gedi tiles, if false will only return file paths}
}
\value{
vector names of filepaths and downloads gedi tiles if download == TRUE
}
\description{
Download cleaned tiles of NASA GEDI L2b data. Tiles are indexed by the lower left corner of the quadrant.
Each tile is 1x1 degree. Cleaning of GEDI data was done following methods in
Burns et al. 2024. Multi-resolution gredded maps of vegetation structure from GEDI
https://doi.org/10.1038/s41597-024-03668-4
These points all have leaf on conditions
}
