% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_modis_lc}
\alias{get_modis_lc}
\title{Download MODIS data}
\usage{
get_modis_lc(
  e,
  start,
  end,
  usr,
  pwd,
  collection = "MCD12Q1",
  variables = c("LC_Type1", "QC"),
  regions,
  outpath
)
}
\arguments{
\item{e}{list of spatExtents to download modis landcover data (e.g., list(ext(xmin, xmax, ymin, ymax)))}

\item{start}{character start date (e.g., "2019-01-01")}

\item{end}{character end date}

\item{usr}{earthdata username}

\item{pwd}{earthdata password}

\item{collection}{collection to download (default MCD12Q1.061) Can be any collection supported by modisfast package}

\item{variables}{vector of variables to download (must be variables in the collection)}

\item{regions}{character name of regions for which data will be downloaded (this will be used to name folders for downloaded files)}

\item{outpath}{file path to save downloads}
}
\value{
dataframe describing downloaded data. See documentation of modisfast::mf_download_data for variable descriptions
Data will be downloaded and combined into a single netcds file
}
\description{
Downloads MODIS data using modisfast R package. Defaults are to download MODIS landcover
}
