% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_soil}
\alias{get_soil}
\title{get_soil
Download sand, silt, and clay content at 0-5 cm depth from soilgrids database
This is an older version of the function (pre-microclimdata package)
Preference is to use get_soil2}
\usage{
get_soil(e, outdir, overwrite = FALSE)
}
\arguments{
\item{e}{spatvector giving extent over which to get soil data}

\item{outdir}{filename to save soil data (ending in .tif)}

\item{overwrite}{whether to overwrite file is it exists (TRUE) or not (default FALSE)}
}
\value{
saves spatraster to outdir and returns spatraster
}
\description{
get_soil
Download sand, silt, and clay content at 0-5 cm depth from soilgrids database
This is an older version of the function (pre-microclimdata package)
Preference is to use get_soil2
}
