% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_soil2}
\alias{get_soil2}
\title{get_soil2
Download sand, silt, and clay content at 0-5 cm depth from soilgrids database using microclimdata package}
\usage{
get_soil2(
  x,
  coords = c("x", "y"),
  crs = "epsg:4326",
  tempdir_location,
  fout,
  overwrite = T
)
}
\arguments{
\item{x}{spatVector, list of spatrasters, or dataframe of coordinates}

\item{coords}{if providing a dataframe, the column names for x and y coordinates}

\item{crs}{if providing a dataframe, the crs}

\item{tempdir_location}{directory to create tempdir folders}

\item{fout}{directory to story output}

\item{overwrite}{whether to overwrite file is it exists (TRUE) or not (default FALSE)}
}
\value{
saves spatraster to outdir and returns spatraster
}
\description{
get_soil2
Download sand, silt, and clay content at 0-5 cm depth from soilgrids database using microclimdata package
}
