% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_topo}
\alias{get_topo}
\title{Get topographic data}
\usage{
get_topo(r, download, outdir)
}
\arguments{
\item{download}{logical; if TRUE, downloads elevation data from amazon web service
and r should be a template raster.
If FALSE, r should be a DEM and slope and aspect will be calculated from r}

\item{outdir}{output directory to store the topography data.}

\item{e}{spatvector giving extent over which to get soil data}

\item{overwrite}{whether to overwrite file is it exists (TRUE) or not (default FALSE)}
}
\value{
saves spatraster to outdir and returns spatraster

returns topography data
}
\description{
Download or provide elevation data. Then calculate slope and aspect.
}
\details{
get_soil
Download sand, silt, and clay content at 0-5 cm depth from soilgrids database
This is an older version of the function (pre-microclimdata package)
Preference is to use get_soil2
}
