% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{getera5tiles}
\alias{getera5tiles}
\title{get splined era5 tile paths from Dave's downloads}
\usage{
getera5tiles(e, tilepath, year)
}
\arguments{
\item{e}{spatExtent}

\item{tilepath}{path to splined era5 tiles}

\item{year}{year of climate data}
}
\description{
get splined era5 tile paths from Dave's downloads
}
