% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microclim.R
\name{prep_micropoint}
\alias{prep_micropoint}
\title{Prep_micropoint
prepare input to micropoint model}
\usage{
prep_micropoint(
  e,
  start,
  end,
  era_path,
  landcover_path,
  soilpath,
  elev_path,
  dtmc_path
)
}
\arguments{
\item{e}{spatextent of area to model}

\item{start}{start date character "YYYY-MM-DD"}

\item{end}{end date; character "YYYY-MM-DD"}

\item{era_path}{file path to downloaded era5 data (should be a .nc file from get_era5)}

\item{landcover_path}{path to landcover data}

\item{soilpath}{path to soil data}

\item{elev_path}{path to elevation data}

\item{dtmc_path}{path to era5 elevation raster}
}
\description{
Prep_micropoint
prepare input to micropoint model
}
