% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{reclass_landcover}
\alias{reclass_landcover}
\title{reclass_landcover
Reclassify landcover types to those specified in microclimf}
\usage{
reclass_landcover(lc, type = "modis_igbp")
}
\arguments{
\item{lc}{spatraster with numeric landcover classifications}

\item{type}{character classification scheme (currently supported is LC_Type 1
(annual international geosphere-biosphere programme classification) from
MODIS MCD12Q1.)}
}
\description{
type options: modis_igbp (IGBP classification from MODIS12Q1)
}
