% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_microclim.R
\name{run_micropoint}
\alias{run_micropoint}
\title{Run micropoint
Iteratively run micropoint model for all points in gedi}
\usage{
run_micropoint(
  tme,
  gedi,
  climr,
  vegp,
  soil,
  elev,
  asp,
  slp,
  dtmc,
  method,
  plotout,
  n,
  maxiter,
  fout = NA,
  modis_path,
  reqhgts = NA,
  vertpai_method = "pai"
)
}
\arguments{
\item{tme}{vector of POSIXlt representing times of climdata}

\item{gedi}{datatable of gedi shots to model microclimates}

\item{climr}{climate rasters - output of mcera5::extract_clima}

\item{vegp}{output of microclimf::vegpfromhab}

\item{soil}{3 layer soil raster with sand, silt, and clay from soilgrids database}

\item{elev}{elevation raster}

\item{asp}{aspect raster}

\item{slp}{slope raster}

\item{dtmc}{raster of elevations of era5 climate data (in climr)}

\item{method}{either "temporal_year", "temporal_month, or "vertprof";
temporal_year models microclimate over entire year at specified reqhgt and estimates monthly variation in PAI based on MODIS LAI data.
The estimate of monthly variation in based on offsets from MODIS in the month when the gedi shot was taken and assumes a constant offset throughout the year.
temporal_month models hourly microclimates at specified reqhgts for the month when the gedi shot was taken. If reqhgt = NA, then
the reqhgts are taken to be 0.15, 2, and then 5 to floor(top of canopy) at 5 m intervals.
vertprof models vertical microclimate profile at specified hour and for specified climate variable over n canopy layers.
There are still some minor bugs with temporal_year and vertprof. The estimate of monthly variation in vertical profiles is unlikely to be reliable
because greenup happens at different times in the understory vs. in the canopy. Vertprof ends up estimating microclimate above the height given to the model.
I'm not sure why this is happening.}

\item{plotout}{one of "tair", "relhum", or "tleaf" - the climate variable to be output for the vertical profile model
only needed if method = "vertprof"}

\item{n}{the number of canopy nodes used to estimate the vertical profile
only needed if method = "vertprof"}

\item{fout}{.csv file to save model outputs}

\item{modis_path}{file path to modis lai data}

\item{reqhgts}{vector of hgts above ground at which to model microclimates.
default NA models microclimate at 0.15, 2, and then 5 to floor(canopy height) at 5 m intervals}

\item{vertpai_method}{determines the method used to calculate the vertical distribution of foliage.
Can be "pai" or "pavd". If "pai", then vertical foliage profiles are calculated
by finding the difference between cumulative PAI at different heights in the canopy. For example,
the pai in the 0-1m voxel would be calculated as the difference between the pai at 0m (cumulative pai)
and the pai at 1m (where this represents the PAI from the top of the canopy down to 1m). If "pavd", then
vertical pai is calculated as proportion of total pai where proportions are determined by the contribution
of pavd in a given layer to the total pavd. These two methods will return slightly different vertical profiles.}
}
\description{
Run micropoint
Iteratively run micropoint model for all points in gedi
}
