% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{vegp_point}
\alias{vegp_point}
\title{Get point vegetation data
Get vegetation data at point based on habitat variables and manually specified height and pai}
\usage{
vegp_point(vegp, lon, lat, pai, h)
}
\description{
Get point vegetation data
Get vegetation data at point based on habitat variables and manually specified height and pai
}
