% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{get_climate}
\alias{get_climate}
\title{Download and process climate data
use functions from microclimdata package}
\usage{
get_climate(
  source,
  r,
  tme,
  creds,
  out = "grid",
  rawout = NA,
  processout,
  file_prefix,
  clean = F
)
}
\arguments{
\item{source}{character; source to get climate data. Accepts
era5
era5-hipergator: this gets data from Dave's splined tiles on hipergator}

\item{r}{a SpatRaster object used as a template for downloading data with an EPSG code defined. Should be in a projected crs.}

\item{tme}{POSIXct vector of hourly time intervals to download data (must be in UTC)}

\item{creds}{dataframe with credentials as specified in microclimdata}

\item{out}{one of 'grid' or 'point' specifying whether to return a list of arrays of climate data or a data.frame for a single point location.}

\item{rawout}{directory to store raw downloaded data. If NA, a temporary directory will be created and deleted after running}

\item{file_prefix}{prefix used for file name}

\item{clean}{logical; if T, deletes original downloads, if F, keeps original downloads}

\item{microout}{directory to store microclimate data grids processed for use in the models}
}
\description{
Download and process climate data
use functions from microclimdata package
}
