% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_params}
\alias{soil_params}
\title{soil_params
get soil parameters for microclimate model}
\usage{
soil_params(soilr)
}
\arguments{
\item{soilr}{spatraster of soildata with three layers named sand, silt, and clay}
}
\value{
spatraster of soiltypes numbered according to microclimf::soilparams
}
\description{
soil_params
get soil parameters for microclimate model
}
